Move ThemeSwitcher into components folder
